Modified MAXINPUTSIZE, made end a global variable.
The maximum input size is 126, since we must allow enough room for the two byte header.
However, we still want to malloc an array of size 128.

Since we will call allocate multiple time, it doesn't make sense to recompute the end
address each time. For this reason, I made end a pointer to the end of the array